Move StorageManager to storage package
